feat: Using channels instead of queue
Now the program uses a standard go channel instead of a queue.
Tests have revealed that a buffer size of 5000 allows to reach
optimal results. The median execution time of the two implementations
differs for less than a millisecond. The usage of a fixed buffer
size also allows to avoid the risk of using too much memory in
extreme cases.